Fix by adding and testing using createApp
