+++ test/integration/test2.js
@@ -2,12 +2,15 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 const { expect } = chai;
 
+const auth = require('../../auth.js');
+const createAppWithStub = () => require('../../createApp.js')(auth);
+
 chai.use(chaiHttp);
 let server;
 let agent;
 describe('route /allowUser2', () => {
   before(async () => {
-    server = require('../../app.js')();
+    server = createAppWithStub();
     agent = chai.request.agent(server.app);
   });
